Ovis/requirements.txt
torch==2.4.0
transformers==4.46.2
tokenizers==0.20.3
sentencepiece==0.1.99
pyarrow==18.0.0
accelerate==1.1.0
pydantic==2.8.2
markdown2[all]
numpy==1.24.3
scikit-learn==1.2.2
requests
httpx
uvicorn
fastapi==0.112.4
einops==0.6.1
einops-exts==0.0.4
timm==0.6.13
tiktoken
transformers_stream_generator==0.0.4
scipy
pandas
torchaudio
xformers
pillow==10.3.0
deepspeed==0.15.4
pysubs2==1.7.2
trl==0.12.1
moviepy
gradio